+++ src/open_refinery/__init__.py
@@ -165,7 +165,7 @@ from .users import (
     valid_role,
 )
 
-__version__ = "1.13.0"
+__version__ = "1.14.0"
 
 __all__ = [
     "Factory",